Aceptar negativos
Ahora acepta negativos

+++ convertirEntero.h
@@ -1,7 +1,7 @@
 #ifndef CONVERTIRENTERO_H_INCLUDED
 #define CONVERTIRENTERO_H_INCLUDED
 
-float * decimalOBaseT10Base(char * n, int * Obase, int * vervose);
+double * decimalOBaseT10Base(char * n, int * Obase, int * vervose);
 
 char * decimal10BaseTDBase(double * n, int * DBase, int * vervose);
